+++ tests/test_runner.nim
@@ -2,3 +2,4 @@ import unittest
 import test_template
 import test_context
 import test_glob
+import test_diff